+++ lib/src/base/types/Static.vue
@@ -2,6 +2,17 @@
   <div class="text-center">
     <div class="element-type">
       <h3>{{ formatName(props.name, " / ") }}</h3>
+      <div v-if="instance" class="autoInsertRow">
+        <div>
+          {{ $t('autoInsertInstance') }}
+        </div>
+        <input v-model="autoInsert" type="checkbox" class="autoInsertInput" />
+      </div>
+      <div v-if="instance && pages.length > 0">
+        <div class="pagesReferenceWrapper">
+          <div class="pagesReference">{{ $t('referencedSection', {pages: pages.join(', ')}) }}</div>
+        </div>
+      </div>
       <form>
         <div>
           <subType :name="props.name" @addStatic="addStatic" ref="viewSaved" :locales="locales" :translation-component-support="translationComponentSupport" :sections-user-id="sectionsUserId">
@@ -17,7 +28,7 @@
 
 <script>
 import subType from "../SubTypes/subType.vue";
-import { formatName, importComp } from "../../utils";
+import { formatName, importComp, sectionHeader } from "../../utils";
 
 export default {
   components: {
@@ -45,6 +56,10 @@ export default {
     sectionsUserId: {
       type: String,
       default: ''
+    },
+    instance: {
+      type: Boolean,
+      default: false
     }
   },
   data() {
@@ -52,6 +67,8 @@ export default {
       myHtml: "",
       elements: [],
       imported: false,
+      autoInsert: false,
+      pages: []
     };
   },
   computed: {
@@ -72,7 +89,6 @@ export default {
       return "null";
     },
   },
-
   mounted() {
     if (this.savedView && this.savedView.settings) {
       setTimeout(() => {
@@ -84,6 +100,9 @@ export default {
     setTimeout(() => {
       this.elements = this.$refs.importedComponent.fields;
     }, 10);
+    if (this.props.linked_to !== '') {
+      this.getGlobalType()
+    }
   },
   methods: {
     formatName,
@@ -94,8 +113,32 @@ export default {
         settings,
         id: this.id,
         weight: this.weight,
+        auto_insertion: this.autoInsert
       });
     },
+    getGlobalType() {
+      const token = this.$cookies.get("sections-auth-token");
+      const config = {
+        headers: sectionHeader({ token }),
+      };
+      const URL =
+          `${this.$sections.serverUrl}/project/${this.$sections.projectId}/global-instances/${this.props.name}`;
+      this.$emit("load", true);
+
+      this.$axios.get(URL, config).then((res) => {
+        if (res && res.data) {
+          this.$emit("load", false);
+          this.autoInsert = res.data.auto_insertion
+          if (res.data.pages && res.data.pages.length > 0) {
+            this.pages = res.data.pages.map(p => p.path)
+          }
+        }
+      })
+          .catch((error) => {
+            this.$emit("load", false);
+            this.showToast("Error", "error", error.response.data.message, error.response.data.options);
+          });
+    }
   },
 };
 </script>
@@ -113,5 +156,26 @@ export default {
 }
 h3 {
   font-size: 29px;
+}
+.autoInsertRow {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: 8px;
+  align-items: center;
+}
+.autoInsertInput {
+  width: 15px;
+  height: 15px;
+}
+.pagesReference {
+  font-size: 14px;
+  color: red;
+  width: 500px;
+}
+.pagesReferenceWrapper {
+  width: 100%;
+  display: flex;
+  place-content: center;
 }
 </style>